Moved anims to functions.ts
